proptypes includes, viewIngredients includes
