+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -3,23 +3,9 @@ import PropTypes from 'prop-types';
 
 import Ingredient from './ingredient/ingredient';
 import { Tab } from '@ya.praktikum/react-developer-burger-ui-components'
+import { burgerIngredientsProptypes } from '../../utils/propTypes';
 
 import styles from './burger-ingredients.module.css'
-
-const burgerIngredientsProptypes = PropTypes.shape({
-    _id: PropTypes.string,
-    name: PropTypes.string,
-    type: PropTypes.string,
-    proteins: PropTypes.number,
-    fat: PropTypes.number,
-    carbohydrates: PropTypes.number,
-    calories: PropTypes.number,
-    price: PropTypes.number,
-    image: PropTypes.string,
-    image_mobile: PropTypes.string,
-    image_large: PropTypes.string,
-    __v: PropTypes.number,
-})
 
 const BurgerIngredients = ({data}) => {
     const [tab, setTab] = useState('one');
